requirements.txt
Django==1.10.1
djangorestframework
markdown
django-filter
django-cors-middleware
django-haystack
whoosh
